Add Config Server, Oauth2 Resource server(Keycloak) with SPI kafka-keycloak

+++ services/ProductService/src/main/java/by/cryptic/productservice/service/command/product/handler/ProductDeleteCommandHandler.java
@@ -1,9 +1,11 @@
 package by.cryptic.productservice.service.command.product.handler;
 
-import by.cryptic.utils.event.product.ProductDeletedEvent;
+import by.cryptic.productservice.model.write.Product;
 import by.cryptic.productservice.repository.write.ProductRepository;
 import by.cryptic.productservice.service.command.product.ProductDeleteCommand;
 import by.cryptic.utils.CommandHandler;
+import by.cryptic.utils.event.product.ProductDeletedEvent;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.CacheEvict;
@@ -22,6 +24,12 @@ public class ProductDeleteCommandHandler implements CommandHandler<ProductDelete
     @Transactional
     @CacheEvict(key = "'product:' + #command.productId()")
     public void handle(ProductDeleteCommand command) {
+        Product product = productRepository.findById(command.productId())
+                        .orElseThrow(() -> new EntityNotFoundException("Product with id %s not found"
+                                .formatted(command.productId())));
+        if (!product.getCreatedBy().equals(command.userId())) {
+            throw new RuntimeException("You cannot delete product because it's not yours");
+        }
         productRepository.deleteById(command.productId());
         eventPublisher.publishEvent(new ProductDeletedEvent(command.productId()));
     }
